Replace formatHours() with API's pre-rendered hours.
The LibCal API now provides pre-rendered hours, so we're going to use
that instead of stitching the string together ourselves. (The way we
were doing it was missing a few exception-related cases anyway.)

We're also splitting sets of multiple hours onto separate lines, so that
they don't wrap and break our table.

The small CSS update aligns dates to the top of their row, so they stay
put when the hours take up more than one line.